+++ lib/bmg/operator.rb
@@ -10,6 +10,7 @@ end
 require_relative 'operator/allbut'
 require_relative 'operator/autosummarize'
 require_relative 'operator/autowrap'
+require_relative 'operator/extend'
 require_relative 'operator/project'
 require_relative 'operator/rename'
-require_relative 'operator/extend'
+require_relative 'operator/union'