+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.0 - March 5st, 2018
+
+* Add the Union operator: union returns both the tuples from left and right operands, but
+  strips the duplicates, if any.
+
 # 0.2.0 - January 13st, 2018
 
 * Add the Extend operator: extends operand tuples with attributes resulting from specified